# examples/basic/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Basic API",
    "version": "1.0.0",
    "description": "Example app using Kanji Framework"
  },
  "paths": {
...
    "/users": {
      "post": {
        "summary": "create endpoint",
        "tags": [
          "Users"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "description": "Unique identifier for the user"
                    },
                    "email": {
                      "type": "string",
                      "format": "email",
                      "description": "Email address"
                    },
                    "name": {
                      "type": "string",
                      "description": "Full name"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "inactive",
                        "pending"
                      ],
                      "description": "The status of the user account"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Timestamp when the user was created"
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "name"
                  ]
                }
              }
            }
          }
        },
        "operationId": "createUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The user's primary email address"
                  },
                  "name": {
                    "type": "string",
                    "minItems": 2,
                    "description": "The user's full name"
                  }
                },
                "required": [
                  "email",
                  "name"
                ]
              }
            }
          }
        }
      },
      "get": {
        "summary": "findAll endpoint",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Unique identifier for the user"
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Email address"
                      },
                      "name": {
                        "type": "string",
                        "description": "Full name"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "active",
                          "inactive",
                          "pending"
                        ],
                        "description": "The status of the user account"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Timestamp when the user was created"
                      }
                    },
                    "required": [
                      "id",
                      "email",
                      "name"
                    ]
                  }
                }
              }
            }
          }
        },
        "deprecated": true,
        "operationId": "findAllUsers"
      }
    },
    "/users/me": {
      "get": {
        "summary": "getMe endpoint",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "getMeProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}